Create SumofEvenNumbersAfterQueries.js
Problem was to input a list of numbers, as well as a list of operations to perform on them, and to output a list of the sum of even numbers after each operation.
Main idea was to keep track of the current even sum, and then make different checks to see how a new operator would affect the sum. if both the old int and the new one (after the operation) were even, you simply sub the new value in for the old in the sum, if one was odd, you either added the val to the sum or completely took away the old even val. if both were odd, you did nothing.

Started this problem out with a forcing method, by summing every time, then refactored the code to make it faster by keeping track instead